Move IDrawable to its own package
