+++ src/games/strategy/triplea/ui/screen/SmallMapImageManager.java
@@ -15,6 +15,7 @@ import java.util.logging.Logger;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.Territory;
 import games.strategy.triplea.ui.mapdata.MapData;
+import games.strategy.triplea.ui.screen.drawable.LandTerritoryDrawable;
 import games.strategy.triplea.util.Stopwatch;
 import games.strategy.ui.ImageScrollerSmallView;
 import games.strategy.ui.Util;
