+++ src/games/strategy/triplea/ui/screen/TerritoryOverLayDrawable.java
@@ -11,6 +11,7 @@ import java.util.List;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.Territory;
 import games.strategy.triplea.ui.mapdata.MapData;
+import games.strategy.triplea.ui.screen.drawable.IDrawable;
 
 public class TerritoryOverLayDrawable implements IDrawable {
   public static enum OP {
