+++ src/games/strategy/triplea/ui/screen/TileManager.java
@@ -33,8 +33,25 @@ import games.strategy.triplea.attachments.TerritoryAttachment;
 import games.strategy.triplea.delegate.TerritoryEffectHelper;
 import games.strategy.triplea.ui.IUIContext;
 import games.strategy.triplea.ui.mapdata.MapData;
-import games.strategy.triplea.ui.screen.IDrawable.OptionalExtraBorderLevel;
+import games.strategy.triplea.ui.screen.drawable.BaseMapDrawable;
+import games.strategy.triplea.ui.screen.drawable.BattleDrawable;
+import games.strategy.triplea.ui.screen.drawable.BlockadeZoneDrawable;
+import games.strategy.triplea.ui.screen.drawable.CapitolMarkerDrawable;
+import games.strategy.triplea.ui.screen.drawable.ConvoyZoneDrawable;
+import games.strategy.triplea.ui.screen.drawable.DecoratorDrawable;
+import games.strategy.triplea.ui.screen.drawable.DrawableComparator;
+import games.strategy.triplea.ui.screen.drawable.IDrawable;
+import games.strategy.triplea.ui.screen.drawable.IDrawable.OptionalExtraBorderLevel;
 import games.strategy.triplea.ui.screen.TerritoryOverLayDrawable.OP;
+import games.strategy.triplea.ui.screen.drawable.KamikazeZoneDrawable;
+import games.strategy.triplea.ui.screen.drawable.LandTerritoryDrawable;
+import games.strategy.triplea.ui.screen.drawable.MapTileDrawable;
+import games.strategy.triplea.ui.screen.drawable.OptionalExtraTerritoryBordersDrawable;
+import games.strategy.triplea.ui.screen.drawable.ReliefMapDrawable;
+import games.strategy.triplea.ui.screen.drawable.SeaZoneOutlineDrawable;
+import games.strategy.triplea.ui.screen.drawable.TerritoryEffectDrawable;
+import games.strategy.triplea.ui.screen.drawable.TerritoryNameDrawable;
+import games.strategy.triplea.ui.screen.drawable.VCDrawable;
 import games.strategy.triplea.util.UnitCategory;
 import games.strategy.triplea.util.UnitSeperator;
 import games.strategy.ui.Util;
